New rock punch in. Also uses animation to trigger.
The animation is triggered by "using" the ability, then runs animation to a set time in the animation, before it notifies to start the punch. Punch runs through the full animation before resetting.
This makes it able to spam, without breaking the code or resetting animation. The animation allways has to run fully before breaking.