Small changes to Water
Did a few small changes to water, according to the meeting the 18th of Feb. 2019.
This includes, making the counter push the user backwards, instead of forwards and making the main ability slower in anticipation of the visuals being changed to a "wave" and later the damage to represent this. The damage remains flat, but should follow a curve, ramping up to a max and then falling off for the remainder of the ability